requirements.txt
pymongo==4.6.1
google-generativeai>=0.8.0
python-dotenv==1.0.0
openpyxl==3.1.2
setuptools>=65.0.0
